Refactor: Move search folder into forms folder
